Use newest Hasura
