+++ docker/docker-compose.custom.yml
@@ -43,6 +43,6 @@ services:
 
   jore4-hasura:
     # pin compatible version of jore4 data model
-    image: "hsldevcom/jore4-hasura:hsl-fix-insert_scheduled_stop_point_with_vehicle_mode--20221102-b9340eb12b9d73480e7b7bf8d7d19f7c4c0f4e1d"
+    image: "hsldevcom/jore4-hasura:hsl-main--20221214-11da6ad81ee45990ced0885b1d1d43833711be35"
 networks:
     jore4:
