+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/DefaultCatalog.java
@@ -4,7 +4,6 @@
 package fi.hsl.jore.jore4.jooq;
 
 
-import fi.hsl.jore.jore4.jooq.deleted.Deleted;
 import fi.hsl.jore.jore4.jooq.infrastructure_network.InfrastructureNetwork;
 import fi.hsl.jore.jore4.jooq.internal_service_pattern.InternalServicePattern;
 import fi.hsl.jore.jore4.jooq.journey_pattern.JourneyPattern;
@@ -31,11 +30,6 @@ public class DefaultCatalog extends CatalogImpl {
      * The reference instance of <code>DEFAULT_CATALOG</code>
      */
     public static final DefaultCatalog DEFAULT_CATALOG = new DefaultCatalog();
-
-    /**
-     * The schema <code>deleted</code>.
-     */
-    public final Deleted DELETED = Deleted.DELETED;
 
     /**
      * The schema <code>infrastructure_network</code>.
@@ -77,7 +71,6 @@ public class DefaultCatalog extends CatalogImpl {
     @Override
     public final List<Schema> getSchemas() {
         return Arrays.asList(
-            Deleted.DELETED,
             InfrastructureNetwork.INFRASTRUCTURE_NETWORK,
             InternalServicePattern.INTERNAL_SERVICE_PATTERN,
             JourneyPattern.JOURNEY_PATTERN,
