--- src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/deleted/Routines.java
@@ -1,69 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package fi.hsl.jore.jore4.jooq.deleted;
-
-
-import fi.hsl.jore.jore4.jooq.deleted.routines.RouteShape_1660903023968;
-
-import org.jooq.Configuration;
-import org.jooq.Field;
-import org.jooq.Record;
-
-
-/**
- * Convenience access to all stored procedures and functions in deleted.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Routines {
-
-    /**
-     * @deprecated Unknown data type. Please define an explicit {@link
-     * org.jooq.Binding} to specify how this type should be handled. Deprecation
-     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
-     * code generator configuration.
-     */
-    @Deprecated
-    public static Object routeShape_1660903023968(
-          Configuration configuration
-        , Record routeRow
-    ) {
-        RouteShape_1660903023968 f = new RouteShape_1660903023968();
-        f.setRouteRow(routeRow);
-
-        f.execute(configuration);
-        return f.getReturnValue();
-    }
-
-    /**
-     * @deprecated Unknown data type. Please define an explicit {@link
-     * org.jooq.Binding} to specify how this type should be handled. Deprecation
-     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
-     * code generator configuration.
-     */
-    @Deprecated
-    public static Field<Object> routeShape_1660903023968(
-          Record routeRow
-    ) {
-        RouteShape_1660903023968 f = new RouteShape_1660903023968();
-        f.setRouteRow(routeRow);
-
-        return f.asField();
-    }
-
-    /**
-     * @deprecated Unknown data type. Please define an explicit {@link
-     * org.jooq.Binding} to specify how this type should be handled. Deprecation
-     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
-     * code generator configuration.
-     */
-    @Deprecated
-    public static Field<Object> routeShape_1660903023968(
-          Field<Record> routeRow
-    ) {
-        RouteShape_1660903023968 f = new RouteShape_1660903023968();
-        f.setRouteRow(routeRow);
-
-        return f.asField();
-    }
-}
