--- src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/internal_service_pattern/Tables.java
@@ -1,136 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package fi.hsl.jore.jore4.jooq.internal_service_pattern;
-
-
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.tables.GetScheduledStopPointsWithNew;
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.tables.ScheduledStopPoint;
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.tables.ScheduledStopPointInvariant;
-
-import java.time.LocalDate;
-import java.util.UUID;
-
-import org.jooq.Configuration;
-import org.jooq.Field;
-import org.jooq.Record;
-import org.jooq.Result;
-
-
-/**
- * Convenience access to all tables in internal_service_pattern.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class Tables {
-
-    /**
-     * The table
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new</code>.
-     */
-    public static final GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW = GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW;
-
-    /**
-     * @deprecated Unknown data type. Please define an explicit {@link
-     * org.jooq.Binding} to specify how this type should be handled. Deprecation
-     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
-     * code generator configuration.
-     */
-    @Deprecated
-    public static Result<Record> GET_SCHEDULED_STOP_POINTS_WITH_NEW(
-          Configuration configuration
-        , UUID replaceScheduledStopPointId
-        , UUID newScheduledStopPointId
-        , UUID newLocatedOnInfrastructureLinkId
-        , Object newMeasuredLocation
-        , String newDirection
-        , String newLabel
-        , LocalDate newValidityStart
-        , LocalDate newValidityEnd
-        , Integer newPriority
-    ) {
-        return configuration.dsl().selectFrom(fi.hsl.jore.jore4.jooq.internal_service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
-              replaceScheduledStopPointId
-            , newScheduledStopPointId
-            , newLocatedOnInfrastructureLinkId
-            , newMeasuredLocation
-            , newDirection
-            , newLabel
-            , newValidityStart
-            , newValidityEnd
-            , newPriority
-        )).fetch();
-    }
-
-    /**
-     * @deprecated Unknown data type. Please define an explicit {@link
-     * org.jooq.Binding} to specify how this type should be handled. Deprecation
-     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
-     * code generator configuration.
-     */
-    @Deprecated
-    public static GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW(
-          UUID replaceScheduledStopPointId
-        , UUID newScheduledStopPointId
-        , UUID newLocatedOnInfrastructureLinkId
-        , Object newMeasuredLocation
-        , String newDirection
-        , String newLabel
-        , LocalDate newValidityStart
-        , LocalDate newValidityEnd
-        , Integer newPriority
-    ) {
-        return fi.hsl.jore.jore4.jooq.internal_service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
-            replaceScheduledStopPointId,
-            newScheduledStopPointId,
-            newLocatedOnInfrastructureLinkId,
-            newMeasuredLocation,
-            newDirection,
-            newLabel,
-            newValidityStart,
-            newValidityEnd,
-            newPriority
-        );
-    }
-
-    /**
-     * @deprecated Unknown data type. Please define an explicit {@link
-     * org.jooq.Binding} to specify how this type should be handled. Deprecation
-     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
-     * code generator configuration.
-     */
-    @Deprecated
-    public static GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW(
-          Field<UUID> replaceScheduledStopPointId
-        , Field<UUID> newScheduledStopPointId
-        , Field<UUID> newLocatedOnInfrastructureLinkId
-        , Field<Object> newMeasuredLocation
-        , Field<String> newDirection
-        , Field<String> newLabel
-        , Field<LocalDate> newValidityStart
-        , Field<LocalDate> newValidityEnd
-        , Field<Integer> newPriority
-    ) {
-        return fi.hsl.jore.jore4.jooq.internal_service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
-            replaceScheduledStopPointId,
-            newScheduledStopPointId,
-            newLocatedOnInfrastructureLinkId,
-            newMeasuredLocation,
-            newDirection,
-            newLabel,
-            newValidityStart,
-            newValidityEnd,
-            newPriority
-        );
-    }
-
-    /**
-     * The table <code>internal_service_pattern.scheduled_stop_point</code>.
-     */
-    public static final ScheduledStopPoint SCHEDULED_STOP_POINT = ScheduledStopPoint.SCHEDULED_STOP_POINT;
-
-    /**
-     * The table
-     * <code>internal_service_pattern.scheduled_stop_point_invariant</code>.
-     */
-    public static final ScheduledStopPointInvariant SCHEDULED_STOP_POINT_INVARIANT = ScheduledStopPointInvariant.SCHEDULED_STOP_POINT_INVARIANT;
-}
