--- src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/internal_service_pattern/tables/ScheduledStopPoint.java
@@ -1,167 +1,0 @@
-/*
- * This file is generated by jOOQ.
- */
-package fi.hsl.jore.jore4.jooq.internal_service_pattern.tables;
-
-
-import fi.hsl.jore.importer.config.jooq.converter.geometry.PointBinding;
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.InternalServicePattern;
-
-import java.time.LocalDate;
-import java.util.UUID;
-
-import org.jooq.Field;
-import org.jooq.ForeignKey;
-import org.jooq.Name;
-import org.jooq.Record;
-import org.jooq.Schema;
-import org.jooq.Table;
-import org.jooq.TableField;
-import org.jooq.TableOptions;
-import org.jooq.impl.DSL;
-import org.jooq.impl.SQLDataType;
-import org.jooq.impl.TableImpl;
-import org.locationtech.jts.geom.Point;
-
-
-/**
- * This class is generated by jOOQ.
- */
-@SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class ScheduledStopPoint extends TableImpl<Record> {
-
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * The reference instance of
-     * <code>internal_service_pattern.scheduled_stop_point</code>
-     */
-    public static final ScheduledStopPoint SCHEDULED_STOP_POINT = new ScheduledStopPoint();
-
-    /**
-     * The class holding records for this type
-     */
-    @Override
-    public Class<Record> getRecordType() {
-        return Record.class;
-    }
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.scheduled_stop_point_id</code>.
-     */
-    public final TableField<Record, UUID> SCHEDULED_STOP_POINT_ID = createField(DSL.name("scheduled_stop_point_id"), SQLDataType.UUID.nullable(false).defaultValue(DSL.field("gen_random_uuid()", SQLDataType.UUID)), this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.measured_location</code>.
-     */
-    public final TableField<Record, Point> MEASURED_LOCATION = createField(DSL.name("measured_location"), org.jooq.impl.DefaultDataType.getDefaultDataType("\"public\".\"geography\"").nullable(false), this, "", new PointBinding());
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.located_on_infrastructure_link_id</code>.
-     */
-    public final TableField<Record, UUID> LOCATED_ON_INFRASTRUCTURE_LINK_ID = createField(DSL.name("located_on_infrastructure_link_id"), SQLDataType.UUID.nullable(false), this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.direction</code>.
-     */
-    public final TableField<Record, String> DIRECTION = createField(DSL.name("direction"), SQLDataType.CLOB.nullable(false), this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.label</code>.
-     */
-    public final TableField<Record, String> LABEL = createField(DSL.name("label"), SQLDataType.CLOB.nullable(false), this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.validity_start</code>.
-     * end of the route's operating date span in the route's local time
-     */
-    public final TableField<Record, LocalDate> VALIDITY_START = createField(DSL.name("validity_start"), SQLDataType.LOCALDATE, this, "end of the route's operating date span in the route's local time");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.validity_end</code>.
-     * end of the operating date span in the scheduled stop point's local time
-     */
-    public final TableField<Record, LocalDate> VALIDITY_END = createField(DSL.name("validity_end"), SQLDataType.LOCALDATE, this, "end of the operating date span in the scheduled stop point's local time");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.scheduled_stop_point.priority</code>.
-     */
-    public final TableField<Record, Integer> PRIORITY = createField(DSL.name("priority"), SQLDataType.INTEGER.nullable(false), this, "");
-
-    private ScheduledStopPoint(Name alias, Table<Record> aliased) {
-        this(alias, aliased, null);
-    }
-
-    private ScheduledStopPoint(Name alias, Table<Record> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table());
-    }
-
-    /**
-     * Create an aliased
-     * <code>internal_service_pattern.scheduled_stop_point</code> table
-     * reference
-     */
-    public ScheduledStopPoint(String alias) {
-        this(DSL.name(alias), SCHEDULED_STOP_POINT);
-    }
-
-    /**
-     * Create an aliased
-     * <code>internal_service_pattern.scheduled_stop_point</code> table
-     * reference
-     */
-    public ScheduledStopPoint(Name alias) {
-        this(alias, SCHEDULED_STOP_POINT);
-    }
-
-    /**
-     * Create a <code>internal_service_pattern.scheduled_stop_point</code> table
-     * reference
-     */
-    public ScheduledStopPoint() {
-        this(DSL.name("scheduled_stop_point"), null);
-    }
-
-    public <O extends Record> ScheduledStopPoint(Table<O> child, ForeignKey<O, Record> key) {
-        super(child, key, SCHEDULED_STOP_POINT);
-    }
-
-    @Override
-    public Schema getSchema() {
-        return aliased() ? null : InternalServicePattern.INTERNAL_SERVICE_PATTERN;
-    }
-
-    @Override
-    public ScheduledStopPoint as(String alias) {
-        return new ScheduledStopPoint(DSL.name(alias), this);
-    }
-
-    @Override
-    public ScheduledStopPoint as(Name alias) {
-        return new ScheduledStopPoint(alias, this);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public ScheduledStopPoint rename(String name) {
-        return new ScheduledStopPoint(DSL.name(name), null);
-    }
-
-    /**
-     * Rename this table
-     */
-    @Override
-    public ScheduledStopPoint rename(Name name) {
-        return new ScheduledStopPoint(name, null);
-    }
-}
