+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/route/Tables.java
@@ -35,7 +35,8 @@ public class Tables {
     public static final Line LINE = Line.LINE;
 
     /**
-     * The table <code>route.route</code>.
+     * The routes from Transmodel:
+     * https://www.transmodel-cen.eu/model/index.htm?goto=2:1:3:483
      */
     public static final Route ROUTE_ = Route.ROUTE_;
