+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/route/tables/Route.java
@@ -22,7 +22,8 @@ import org.jooq.impl.TableImpl;
 
 
 /**
- * This class is generated by jOOQ.
+ * The routes from Transmodel:
+ * https://www.transmodel-cen.eu/model/index.htm?goto=2:1:3:483
  */
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
 public class Route extends TableImpl<Record> {
@@ -43,44 +44,57 @@ public class Route extends TableImpl<Record> {
     }
 
     /**
-     * The column <code>route.route.route_id</code>.
+     * The column <code>route.route.route_id</code>. The ID of the route.
      */
-    public final TableField<Record, UUID> ROUTE_ID = createField(DSL.name("route_id"), SQLDataType.UUID.nullable(false).defaultValue(DSL.field("gen_random_uuid()", SQLDataType.UUID)), this, "");
+    public final TableField<Record, UUID> ROUTE_ID = createField(DSL.name("route_id"), SQLDataType.UUID.nullable(false).defaultValue(DSL.field("gen_random_uuid()", SQLDataType.UUID)), this, "The ID of the route.");
 
     /**
-     * The column <code>route.route.description_i18n</code>.
+     * The column <code>route.route.description_i18n</code>. The description of
+     * the route in the form of starting location - destination. Placeholder for
+     * multilingual strings.
      */
-    public final TableField<Record, JSONB> DESCRIPTION_I18N = createField(DSL.name("description_i18n"), SQLDataType.JSONB, this, "");
+    public final TableField<Record, JSONB> DESCRIPTION_I18N = createField(DSL.name("description_i18n"), SQLDataType.JSONB, this, "The description of the route in the form of starting location - destination. Placeholder for multilingual strings.");
 
     /**
-     * The column <code>route.route.on_line_id</code>.
+     * The column <code>route.route.on_line_id</code>. The line to which this
+     * route belongs.
      */
-    public final TableField<Record, UUID> ON_LINE_ID = createField(DSL.name("on_line_id"), SQLDataType.UUID.nullable(false), this, "");
+    public final TableField<Record, UUID> ON_LINE_ID = createField(DSL.name("on_line_id"), SQLDataType.UUID.nullable(false), this, "The line to which this route belongs.");
 
     /**
-     * The column <code>route.route.validity_start</code>.
+     * The column <code>route.route.validity_start</code>. The point in time
+     * when the route becomes valid. If NULL, the route has been always valid
+     * before end time of validity period.
      */
-    public final TableField<Record, LocalDate> VALIDITY_START = createField(DSL.name("validity_start"), SQLDataType.LOCALDATE, this, "");
+    public final TableField<Record, LocalDate> VALIDITY_START = createField(DSL.name("validity_start"), SQLDataType.LOCALDATE, this, "The point in time when the route becomes valid. If NULL, the route has been always valid before end time of validity period.");
 
     /**
-     * The column <code>route.route.validity_end</code>.
+     * The column <code>route.route.validity_end</code>. The point in time from
+     * which onwards the route is no longer valid. If NULL, the route is valid
+     * indefinitely after the start time of the validity period.
      */
-    public final TableField<Record, LocalDate> VALIDITY_END = createField(DSL.name("validity_end"), SQLDataType.LOCALDATE, this, "");
+    public final TableField<Record, LocalDate> VALIDITY_END = createField(DSL.name("validity_end"), SQLDataType.LOCALDATE, this, "The point in time from which onwards the route is no longer valid. If NULL, the route is valid indefinitely after the start time of the validity period.");
 
     /**
-     * The column <code>route.route.priority</code>.
+     * The column <code>route.route.priority</code>. The priority of the route
+     * definition. The definition may be overridden by higher priority
+     * definitions.
      */
-    public final TableField<Record, Integer> PRIORITY = createField(DSL.name("priority"), SQLDataType.INTEGER.nullable(false), this, "");
+    public final TableField<Record, Integer> PRIORITY = createField(DSL.name("priority"), SQLDataType.INTEGER.nullable(false), this, "The priority of the route definition. The definition may be overridden by higher priority definitions.");
 
     /**
-     * The column <code>route.route.label</code>.
+     * The column <code>route.route.label</code>. The label of the route
+     * definition, label and direction together are unique for a certain
+     * priority and validity period.
      */
-    public final TableField<Record, String> LABEL = createField(DSL.name("label"), SQLDataType.CLOB.nullable(false), this, "");
+    public final TableField<Record, String> LABEL = createField(DSL.name("label"), SQLDataType.CLOB.nullable(false), this, "The label of the route definition, label and direction together are unique for a certain priority and validity period.");
 
     /**
-     * The column <code>route.route.direction</code>.
+     * The column <code>route.route.direction</code>. The direction of the route
+     * definition, label and direction together are unique for a certain
+     * priority and validity period.
      */
-    public final TableField<Record, String> DIRECTION = createField(DSL.name("direction"), SQLDataType.CLOB.nullable(false), this, "");
+    public final TableField<Record, String> DIRECTION = createField(DSL.name("direction"), SQLDataType.CLOB.nullable(false), this, "The direction of the route definition, label and direction together are unique for a certain priority and validity period.");
 
     /**
      * The column <code>route.route.name_i18n</code>.
@@ -112,7 +126,7 @@ public class Route extends TableImpl<Record> {
     }
 
     private Route(Name alias, Table<Record> aliased, Field<?>[] parameters) {
-        super(alias, null, aliased, parameters, DSL.comment(""), TableOptions.table());
+        super(alias, null, aliased, parameters, DSL.comment("The routes from Transmodel: https://www.transmodel-cen.eu/model/index.htm?goto=2:1:3:483"), TableOptions.table());
     }
 
     /**
