+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/Routines.java
@@ -1,0 +1,240 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package fi.hsl.jore.jore4.jooq.service_pattern;
+
+
+import fi.hsl.jore.jore4.jooq.service_pattern.routines.DeleteScheduledStopPointLabel;
+import fi.hsl.jore.jore4.jooq.service_pattern.routines.InsertScheduledStopPointLabel;
+import fi.hsl.jore.jore4.jooq.service_pattern.routines.ScheduledStopPointClosestPointOnInfrastructureLink;
+import fi.hsl.jore.jore4.jooq.service_pattern.routines.ScheduledStopPointRelativeDistanceFromInfrastructureLink;
+import fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew;
+
+import java.time.LocalDate;
+import java.util.UUID;
+
+import org.jooq.Configuration;
+import org.jooq.Field;
+import org.jooq.Record;
+import org.jooq.Result;
+
+
+/**
+ * Convenience access to all stored procedures and functions in service_pattern.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+public class Routines {
+
+    /**
+     * Call <code>service_pattern.delete_scheduled_stop_point_label</code>
+     */
+    public static void deleteScheduledStopPointLabel(
+          Configuration configuration
+        , String oldLabel
+    ) {
+        DeleteScheduledStopPointLabel p = new DeleteScheduledStopPointLabel();
+        p.setOldLabel(oldLabel);
+
+        p.execute(configuration);
+    }
+
+    /**
+     * Call <code>service_pattern.insert_scheduled_stop_point_label</code>
+     */
+    public static void insertScheduledStopPointLabel(
+          Configuration configuration
+        , String newLabel
+    ) {
+        InsertScheduledStopPointLabel p = new InsertScheduledStopPointLabel();
+        p.setNewLabel(newLabel);
+
+        p.execute(configuration);
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static Object scheduledStopPointClosestPointOnInfrastructureLink(
+          Configuration configuration
+        , Record ssp
+    ) {
+        ScheduledStopPointClosestPointOnInfrastructureLink f = new ScheduledStopPointClosestPointOnInfrastructureLink();
+        f.setSsp(ssp);
+
+        f.execute(configuration);
+        return f.getReturnValue();
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static Field<Object> scheduledStopPointClosestPointOnInfrastructureLink(
+          Record ssp
+    ) {
+        ScheduledStopPointClosestPointOnInfrastructureLink f = new ScheduledStopPointClosestPointOnInfrastructureLink();
+        f.setSsp(ssp);
+
+        return f.asField();
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static Field<Object> scheduledStopPointClosestPointOnInfrastructureLink(
+          Field<Record> ssp
+    ) {
+        ScheduledStopPointClosestPointOnInfrastructureLink f = new ScheduledStopPointClosestPointOnInfrastructureLink();
+        f.setSsp(ssp);
+
+        return f.asField();
+    }
+
+    /**
+     * Call
+     * <code>service_pattern.scheduled_stop_point_relative_distance_from_infrastructure_link</code>
+     */
+    public static Double scheduledStopPointRelativeDistanceFromInfrastructureLink(
+          Configuration configuration
+        , Record ssp
+    ) {
+        ScheduledStopPointRelativeDistanceFromInfrastructureLink f = new ScheduledStopPointRelativeDistanceFromInfrastructureLink();
+        f.setSsp(ssp);
+
+        f.execute(configuration);
+        return f.getReturnValue();
+    }
+
+    /**
+     * Get
+     * <code>service_pattern.scheduled_stop_point_relative_distance_from_infrastructure_link</code>
+     * as a field.
+     */
+    public static Field<Double> scheduledStopPointRelativeDistanceFromInfrastructureLink(
+          Record ssp
+    ) {
+        ScheduledStopPointRelativeDistanceFromInfrastructureLink f = new ScheduledStopPointRelativeDistanceFromInfrastructureLink();
+        f.setSsp(ssp);
+
+        return f.asField();
+    }
+
+    /**
+     * Get
+     * <code>service_pattern.scheduled_stop_point_relative_distance_from_infrastructure_link</code>
+     * as a field.
+     */
+    public static Field<Double> scheduledStopPointRelativeDistanceFromInfrastructureLink(
+          Field<Record> ssp
+    ) {
+        ScheduledStopPointRelativeDistanceFromInfrastructureLink f = new ScheduledStopPointRelativeDistanceFromInfrastructureLink();
+        f.setSsp(ssp);
+
+        return f.asField();
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static Result<Record> getScheduledStopPointsWithNew(
+          Configuration configuration
+        , UUID replaceScheduledStopPointId
+        , UUID newScheduledStopPointId
+        , UUID newLocatedOnInfrastructureLinkId
+        , Object newMeasuredLocation
+        , String newDirection
+        , String newLabel
+        , LocalDate newValidityStart
+        , LocalDate newValidityEnd
+        , Integer newPriority
+    ) {
+        return configuration.dsl().selectFrom(fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
+              replaceScheduledStopPointId
+            , newScheduledStopPointId
+            , newLocatedOnInfrastructureLinkId
+            , newMeasuredLocation
+            , newDirection
+            , newLabel
+            , newValidityStart
+            , newValidityEnd
+            , newPriority
+        )).fetch();
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static GetScheduledStopPointsWithNew getScheduledStopPointsWithNew(
+          UUID replaceScheduledStopPointId
+        , UUID newScheduledStopPointId
+        , UUID newLocatedOnInfrastructureLinkId
+        , Object newMeasuredLocation
+        , String newDirection
+        , String newLabel
+        , LocalDate newValidityStart
+        , LocalDate newValidityEnd
+        , Integer newPriority
+    ) {
+        return fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
+            replaceScheduledStopPointId,
+            newScheduledStopPointId,
+            newLocatedOnInfrastructureLinkId,
+            newMeasuredLocation,
+            newDirection,
+            newLabel,
+            newValidityStart,
+            newValidityEnd,
+            newPriority
+        );
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static GetScheduledStopPointsWithNew getScheduledStopPointsWithNew(
+          Field<UUID> replaceScheduledStopPointId
+        , Field<UUID> newScheduledStopPointId
+        , Field<UUID> newLocatedOnInfrastructureLinkId
+        , Field<Object> newMeasuredLocation
+        , Field<String> newDirection
+        , Field<String> newLabel
+        , Field<LocalDate> newValidityStart
+        , Field<LocalDate> newValidityEnd
+        , Field<Integer> newPriority
+    ) {
+        return fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
+            replaceScheduledStopPointId,
+            newScheduledStopPointId,
+            newLocatedOnInfrastructureLinkId,
+            newMeasuredLocation,
+            newDirection,
+            newLabel,
+            newValidityStart,
+            newValidityEnd,
+            newPriority
+        );
+    }
+}
