+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/ServicePattern.java
@@ -5,13 +5,21 @@ package fi.hsl.jore.jore4.jooq.service_pattern;
 
 
 import fi.hsl.jore.jore4.jooq.DefaultCatalog;
+import fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew;
 import fi.hsl.jore.jore4.jooq.service_pattern.tables.ScheduledStopPoint;
+import fi.hsl.jore.jore4.jooq.service_pattern.tables.ScheduledStopPointInvariant;
 import fi.hsl.jore.jore4.jooq.service_pattern.tables.VehicleModeOnScheduledStopPoint;
 
+import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.List;
+import java.util.UUID;
 
 import org.jooq.Catalog;
+import org.jooq.Configuration;
+import org.jooq.Field;
+import org.jooq.Record;
+import org.jooq.Result;
 import org.jooq.Table;
 import org.jooq.impl.SchemaImpl;
 
@@ -30,11 +38,116 @@ public class ServicePattern extends SchemaImpl {
     public static final ServicePattern SERVICE_PATTERN = new ServicePattern();
 
     /**
+     * The table
+     * <code>service_pattern.get_scheduled_stop_points_with_new</code>.
+     */
+    public final GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW = GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW;
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static Result<Record> GET_SCHEDULED_STOP_POINTS_WITH_NEW(
+          Configuration configuration
+        , UUID replaceScheduledStopPointId
+        , UUID newScheduledStopPointId
+        , UUID newLocatedOnInfrastructureLinkId
+        , Object newMeasuredLocation
+        , String newDirection
+        , String newLabel
+        , LocalDate newValidityStart
+        , LocalDate newValidityEnd
+        , Integer newPriority
+    ) {
+        return configuration.dsl().selectFrom(fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
+              replaceScheduledStopPointId
+            , newScheduledStopPointId
+            , newLocatedOnInfrastructureLinkId
+            , newMeasuredLocation
+            , newDirection
+            , newLabel
+            , newValidityStart
+            , newValidityEnd
+            , newPriority
+        )).fetch();
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW(
+          UUID replaceScheduledStopPointId
+        , UUID newScheduledStopPointId
+        , UUID newLocatedOnInfrastructureLinkId
+        , Object newMeasuredLocation
+        , String newDirection
+        , String newLabel
+        , LocalDate newValidityStart
+        , LocalDate newValidityEnd
+        , Integer newPriority
+    ) {
+        return fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
+            replaceScheduledStopPointId,
+            newScheduledStopPointId,
+            newLocatedOnInfrastructureLinkId,
+            newMeasuredLocation,
+            newDirection,
+            newLabel,
+            newValidityStart,
+            newValidityEnd,
+            newPriority
+        );
+    }
+
+    /**
+     * @deprecated Unknown data type. Please define an explicit {@link
+     * org.jooq.Binding} to specify how this type should be handled. Deprecation
+     * can be turned off using {@literal <deprecationOnUnknownTypes/>} in your
+     * code generator configuration.
+     */
+    @Deprecated
+    public static GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW(
+          Field<UUID> replaceScheduledStopPointId
+        , Field<UUID> newScheduledStopPointId
+        , Field<UUID> newLocatedOnInfrastructureLinkId
+        , Field<Object> newMeasuredLocation
+        , Field<String> newDirection
+        , Field<String> newLabel
+        , Field<LocalDate> newValidityStart
+        , Field<LocalDate> newValidityEnd
+        , Field<Integer> newPriority
+    ) {
+        return fi.hsl.jore.jore4.jooq.service_pattern.tables.GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW.call(
+            replaceScheduledStopPointId,
+            newScheduledStopPointId,
+            newLocatedOnInfrastructureLinkId,
+            newMeasuredLocation,
+            newDirection,
+            newLabel,
+            newValidityStart,
+            newValidityEnd,
+            newPriority
+        );
+    }
+
+    /**
      * The scheduled stop points:
      * https://www.transmodel-cen.eu/model/index.htm?goto=2:3:4:845 .
      * Colloquially known as stops from the perspective of timetable planning.
      */
     public final ScheduledStopPoint SCHEDULED_STOP_POINT = ScheduledStopPoint.SCHEDULED_STOP_POINT;
+
+    /**
+     * The table <code>service_pattern.scheduled_stop_point_invariant</code>.
+     */
+    public final ScheduledStopPointInvariant SCHEDULED_STOP_POINT_INVARIANT = ScheduledStopPointInvariant.SCHEDULED_STOP_POINT_INVARIANT;
 
     /**
      * Which scheduled stop points are serviced by which vehicle modes?
@@ -57,7 +170,9 @@ public class ServicePattern extends SchemaImpl {
     @Override
     public final List<Table<?>> getTables() {
         return Arrays.asList(
+            GetScheduledStopPointsWithNew.GET_SCHEDULED_STOP_POINTS_WITH_NEW,
             ScheduledStopPoint.SCHEDULED_STOP_POINT,
+            ScheduledStopPointInvariant.SCHEDULED_STOP_POINT_INVARIANT,
             VehicleModeOnScheduledStopPoint.VEHICLE_MODE_ON_SCHEDULED_STOP_POINT
         );
     }
