+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/routines/DeleteScheduledStopPointLabel.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package fi.hsl.jore.jore4.jooq.internal_service_pattern.routines;
+package fi.hsl.jore.jore4.jooq.service_pattern.routines;
 
 
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.InternalServicePattern;
+import fi.hsl.jore.jore4.jooq.service_pattern.ServicePattern;
 
 import org.jooq.Parameter;
 import org.jooq.impl.AbstractRoutine;
@@ -22,7 +22,7 @@ public class DeleteScheduledStopPointLabel extends AbstractRoutine<java.lang.Voi
 
     /**
      * The parameter
-     * <code>internal_service_pattern.delete_scheduled_stop_point_label.old_label</code>.
+     * <code>service_pattern.delete_scheduled_stop_point_label.old_label</code>.
      */
     public static final Parameter<String> OLD_LABEL = Internal.createParameter("old_label", SQLDataType.CLOB, false, false);
 
@@ -30,7 +30,7 @@ public class DeleteScheduledStopPointLabel extends AbstractRoutine<java.lang.Voi
      * Create a new routine call instance
      */
     public DeleteScheduledStopPointLabel() {
-        super("delete_scheduled_stop_point_label", InternalServicePattern.INTERNAL_SERVICE_PATTERN);
+        super("delete_scheduled_stop_point_label", ServicePattern.SERVICE_PATTERN);
 
         addInParameter(OLD_LABEL);
     }
