+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/routines/InsertScheduledStopPointLabel.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package fi.hsl.jore.jore4.jooq.internal_service_pattern.routines;
+package fi.hsl.jore.jore4.jooq.service_pattern.routines;
 
 
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.InternalServicePattern;
+import fi.hsl.jore.jore4.jooq.service_pattern.ServicePattern;
 
 import org.jooq.Parameter;
 import org.jooq.impl.AbstractRoutine;
@@ -22,7 +22,7 @@ public class InsertScheduledStopPointLabel extends AbstractRoutine<java.lang.Voi
 
     /**
      * The parameter
-     * <code>internal_service_pattern.insert_scheduled_stop_point_label.new_label</code>.
+     * <code>service_pattern.insert_scheduled_stop_point_label.new_label</code>.
      */
     public static final Parameter<String> NEW_LABEL = Internal.createParameter("new_label", SQLDataType.CLOB, false, false);
 
@@ -30,7 +30,7 @@ public class InsertScheduledStopPointLabel extends AbstractRoutine<java.lang.Voi
      * Create a new routine call instance
      */
     public InsertScheduledStopPointLabel() {
-        super("insert_scheduled_stop_point_label", InternalServicePattern.INTERNAL_SERVICE_PATTERN);
+        super("insert_scheduled_stop_point_label", ServicePattern.SERVICE_PATTERN);
 
         addInParameter(NEW_LABEL);
     }
