+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/routines/ScheduledStopPointClosestPointOnInfrastructureLink.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package fi.hsl.jore.jore4.jooq.deleted.routines;
+package fi.hsl.jore.jore4.jooq.service_pattern.routines;
 
 
-import fi.hsl.jore.jore4.jooq.deleted.Deleted;
+import fi.hsl.jore.jore4.jooq.service_pattern.ServicePattern;
 
 import org.jooq.Field;
 import org.jooq.Parameter;
@@ -22,7 +22,7 @@ import org.jooq.impl.Internal;
  */
 @Deprecated
 @SuppressWarnings({ "all", "unchecked", "rawtypes" })
-public class RouteShape_1660903023968 extends AbstractRoutine<Object> {
+public class ScheduledStopPointClosestPointOnInfrastructureLink extends AbstractRoutine<Object> {
 
     private static final long serialVersionUID = 1L;
 
@@ -36,32 +36,33 @@ public class RouteShape_1660903023968 extends AbstractRoutine<Object> {
     public static final Parameter<Object> RETURN_VALUE = Internal.createParameter("RETURN_VALUE", org.jooq.impl.DefaultDataType.getDefaultDataType("\"public\".\"geography\""), false, false);
 
     /**
-     * The parameter <code>deleted.route_shape_1660903023968.route_row</code>.
+     * The parameter
+     * <code>service_pattern.scheduled_stop_point_closest_point_on_infrastructure_link.ssp</code>.
      */
-    public static final Parameter<Record> ROUTE_ROW = Internal.createParameter("route_row", fi.hsl.jore.jore4.jooq.route.tables.Route.ROUTE_.getDataType(), false, false);
+    public static final Parameter<Record> SSP = Internal.createParameter("ssp", fi.hsl.jore.jore4.jooq.service_pattern.tables.ScheduledStopPoint.SCHEDULED_STOP_POINT.getDataType(), false, false);
 
     /**
      * Create a new routine call instance
      */
-    public RouteShape_1660903023968() {
-        super("route_shape_1660903023968", Deleted.DELETED, org.jooq.impl.DefaultDataType.getDefaultDataType("\"public\".\"geography\""));
+    public ScheduledStopPointClosestPointOnInfrastructureLink() {
+        super("scheduled_stop_point_closest_point_on_infrastructure_link", ServicePattern.SERVICE_PATTERN, org.jooq.impl.DefaultDataType.getDefaultDataType("\"public\".\"geography\""));
 
         setReturnParameter(RETURN_VALUE);
-        addInParameter(ROUTE_ROW);
+        addInParameter(SSP);
     }
 
     /**
-     * Set the <code>route_row</code> parameter IN value to the routine
+     * Set the <code>ssp</code> parameter IN value to the routine
      */
-    public void setRouteRow(Record value) {
-        setValue(ROUTE_ROW, value);
+    public void setSsp(Record value) {
+        setValue(SSP, value);
     }
 
     /**
-     * Set the <code>route_row</code> parameter to the function to be used with
-     * a {@link org.jooq.Select} statement
+     * Set the <code>ssp</code> parameter to the function to be used with a
+     * {@link org.jooq.Select} statement
      */
-    public void setRouteRow(Field<Record> field) {
-        setField(ROUTE_ROW, field);
+    public void setSsp(Field<Record> field) {
+        setField(SSP, field);
     }
 }
