+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/routines/ScheduledStopPointRelativeDistanceFromInfrastructureLink.java
@@ -1,0 +1,61 @@
+/*
+ * This file is generated by jOOQ.
+ */
+package fi.hsl.jore.jore4.jooq.service_pattern.routines;
+
+
+import fi.hsl.jore.jore4.jooq.service_pattern.ServicePattern;
+
+import org.jooq.Field;
+import org.jooq.Parameter;
+import org.jooq.Record;
+import org.jooq.impl.AbstractRoutine;
+import org.jooq.impl.Internal;
+import org.jooq.impl.SQLDataType;
+
+
+/**
+ * This class is generated by jOOQ.
+ */
+@SuppressWarnings({ "all", "unchecked", "rawtypes" })
+public class ScheduledStopPointRelativeDistanceFromInfrastructureLink extends AbstractRoutine<Double> {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * The parameter
+     * <code>service_pattern.scheduled_stop_point_relative_distance_from_infrastructure_link.RETURN_VALUE</code>.
+     */
+    public static final Parameter<Double> RETURN_VALUE = Internal.createParameter("RETURN_VALUE", SQLDataType.DOUBLE, false, false);
+
+    /**
+     * The parameter
+     * <code>service_pattern.scheduled_stop_point_relative_distance_from_infrastructure_link.ssp</code>.
+     */
+    public static final Parameter<Record> SSP = Internal.createParameter("ssp", fi.hsl.jore.jore4.jooq.service_pattern.tables.ScheduledStopPoint.SCHEDULED_STOP_POINT.getDataType(), false, false);
+
+    /**
+     * Create a new routine call instance
+     */
+    public ScheduledStopPointRelativeDistanceFromInfrastructureLink() {
+        super("scheduled_stop_point_relative_distance_from_infrastructure_link", ServicePattern.SERVICE_PATTERN, SQLDataType.DOUBLE);
+
+        setReturnParameter(RETURN_VALUE);
+        addInParameter(SSP);
+    }
+
+    /**
+     * Set the <code>ssp</code> parameter IN value to the routine
+     */
+    public void setSsp(Record value) {
+        setValue(SSP, value);
+    }
+
+    /**
+     * Set the <code>ssp</code> parameter to the function to be used with a
+     * {@link org.jooq.Select} statement
+     */
+    public void setSsp(Field<Record> field) {
+        setField(SSP, field);
+    }
+}
