+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/tables/GetScheduledStopPointsWithNew.java
@@ -1,11 +1,11 @@
 /*
  * This file is generated by jOOQ.
  */
-package fi.hsl.jore.jore4.jooq.internal_service_pattern.tables;
+package fi.hsl.jore.jore4.jooq.service_pattern.tables;
 
 
 import fi.hsl.jore.importer.config.jooq.converter.geometry.PointBinding;
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.InternalServicePattern;
+import fi.hsl.jore.jore4.jooq.service_pattern.ServicePattern;
 
 import java.time.LocalDate;
 import java.util.UUID;
@@ -33,7 +33,7 @@ public class GetScheduledStopPointsWithNew extends TableImpl<Record> {
 
     /**
      * The reference instance of
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new</code>
+     * <code>service_pattern.get_scheduled_stop_points_with_new</code>
      */
     public static final GetScheduledStopPointsWithNew GET_SCHEDULED_STOP_POINTS_WITH_NEW = new GetScheduledStopPointsWithNew();
 
@@ -47,37 +47,55 @@ public class GetScheduledStopPointsWithNew extends TableImpl<Record> {
 
     /**
      * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.scheduled_stop_point_id</code>.
+     * <code>service_pattern.get_scheduled_stop_points_with_new.scheduled_stop_point_id</code>.
      */
     public final TableField<Record, UUID> SCHEDULED_STOP_POINT_ID = createField(DSL.name("scheduled_stop_point_id"), SQLDataType.UUID, this, "");
 
     /**
      * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.label</code>.
-     */
-    public final TableField<Record, String> LABEL = createField(DSL.name("label"), SQLDataType.CLOB, this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.measured_location</code>.
+     * <code>service_pattern.get_scheduled_stop_points_with_new.measured_location</code>.
      */
     public final TableField<Record, Point> MEASURED_LOCATION = createField(DSL.name("measured_location"), org.jooq.impl.DefaultDataType.getDefaultDataType("\"public\".\"geography\""), this, "", new PointBinding());
 
     /**
      * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.located_on_infrastructure_link_id</code>.
+     * <code>service_pattern.get_scheduled_stop_points_with_new.located_on_infrastructure_link_id</code>.
      */
     public final TableField<Record, UUID> LOCATED_ON_INFRASTRUCTURE_LINK_ID = createField(DSL.name("located_on_infrastructure_link_id"), SQLDataType.UUID, this, "");
 
     /**
      * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.direction</code>.
+     * <code>service_pattern.get_scheduled_stop_points_with_new.direction</code>.
      */
     public final TableField<Record, String> DIRECTION = createField(DSL.name("direction"), SQLDataType.CLOB, this, "");
 
     /**
      * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.relative_distance_from_infrastructure_link_start</code>.
+     * <code>service_pattern.get_scheduled_stop_points_with_new.label</code>.
+     */
+    public final TableField<Record, String> LABEL = createField(DSL.name("label"), SQLDataType.CLOB, this, "");
+
+    /**
+     * The column
+     * <code>service_pattern.get_scheduled_stop_points_with_new.validity_start</code>.
+     */
+    public final TableField<Record, LocalDate> VALIDITY_START = createField(DSL.name("validity_start"), SQLDataType.LOCALDATE, this, "");
+
+    /**
+     * The column
+     * <code>service_pattern.get_scheduled_stop_points_with_new.validity_end</code>.
+     */
+    public final TableField<Record, LocalDate> VALIDITY_END = createField(DSL.name("validity_end"), SQLDataType.LOCALDATE, this, "");
+
+    /**
+     * The column
+     * <code>service_pattern.get_scheduled_stop_points_with_new.priority</code>.
+     */
+    public final TableField<Record, Integer> PRIORITY = createField(DSL.name("priority"), SQLDataType.INTEGER, this, "");
+
+    /**
+     * The column
+     * <code>service_pattern.get_scheduled_stop_points_with_new.relative_distance_from_infrastructure_link_start</code>.
      */
     public final TableField<Record, Double> RELATIVE_DISTANCE_FROM_INFRASTRUCTURE_LINK_START = createField(DSL.name("relative_distance_from_infrastructure_link_start"), SQLDataType.DOUBLE, this, "");
 
@@ -89,24 +107,6 @@ public class GetScheduledStopPointsWithNew extends TableImpl<Record> {
      */
     @Deprecated
     public final TableField<Record, Object> CLOSEST_POINT_ON_INFRASTRUCTURE_LINK = createField(DSL.name("closest_point_on_infrastructure_link"), org.jooq.impl.DefaultDataType.getDefaultDataType("\"public\".\"geography\""), this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.validity_start</code>.
-     */
-    public final TableField<Record, LocalDate> VALIDITY_START = createField(DSL.name("validity_start"), SQLDataType.LOCALDATE, this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.validity_end</code>.
-     */
-    public final TableField<Record, LocalDate> VALIDITY_END = createField(DSL.name("validity_end"), SQLDataType.LOCALDATE, this, "");
-
-    /**
-     * The column
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new.priority</code>.
-     */
-    public final TableField<Record, Integer> PRIORITY = createField(DSL.name("priority"), SQLDataType.INTEGER, this, "");
 
     private GetScheduledStopPointsWithNew(Name alias, Table<Record> aliased) {
         this(alias, aliased, new Field[] {
@@ -128,8 +128,8 @@ public class GetScheduledStopPointsWithNew extends TableImpl<Record> {
 
     /**
      * Create an aliased
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new</code>
-     * table reference
+     * <code>service_pattern.get_scheduled_stop_points_with_new</code> table
+     * reference
      */
     public GetScheduledStopPointsWithNew(String alias) {
         this(DSL.name(alias), GET_SCHEDULED_STOP_POINTS_WITH_NEW);
@@ -137,16 +137,15 @@ public class GetScheduledStopPointsWithNew extends TableImpl<Record> {
 
     /**
      * Create an aliased
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new</code>
-     * table reference
+     * <code>service_pattern.get_scheduled_stop_points_with_new</code> table
+     * reference
      */
     public GetScheduledStopPointsWithNew(Name alias) {
         this(alias, GET_SCHEDULED_STOP_POINTS_WITH_NEW);
     }
 
     /**
-     * Create a
-     * <code>internal_service_pattern.get_scheduled_stop_points_with_new</code>
+     * Create a <code>service_pattern.get_scheduled_stop_points_with_new</code>
      * table reference
      */
     public GetScheduledStopPointsWithNew() {
@@ -155,7 +154,7 @@ public class GetScheduledStopPointsWithNew extends TableImpl<Record> {
 
     @Override
     public Schema getSchema() {
-        return aliased() ? null : InternalServicePattern.INTERNAL_SERVICE_PATTERN;
+        return aliased() ? null : ServicePattern.SERVICE_PATTERN;
     }
 
     @Override
