+++ src/main/java/generated-sources/jooq/fi/hsl/jore/jore4/jooq/service_pattern/tables/ScheduledStopPointInvariant.java
@@ -1,10 +1,10 @@
 /*
  * This file is generated by jOOQ.
  */
-package fi.hsl.jore.jore4.jooq.internal_service_pattern.tables;
+package fi.hsl.jore.jore4.jooq.service_pattern.tables;
 
 
-import fi.hsl.jore.jore4.jooq.internal_service_pattern.InternalServicePattern;
+import fi.hsl.jore.jore4.jooq.service_pattern.ServicePattern;
 
 import org.jooq.Field;
 import org.jooq.ForeignKey;
@@ -29,7 +29,7 @@ public class ScheduledStopPointInvariant extends TableImpl<Record> {
 
     /**
      * The reference instance of
-     * <code>internal_service_pattern.scheduled_stop_point_invariant</code>
+     * <code>service_pattern.scheduled_stop_point_invariant</code>
      */
     public static final ScheduledStopPointInvariant SCHEDULED_STOP_POINT_INVARIANT = new ScheduledStopPointInvariant();
 
@@ -43,7 +43,7 @@ public class ScheduledStopPointInvariant extends TableImpl<Record> {
 
     /**
      * The column
-     * <code>internal_service_pattern.scheduled_stop_point_invariant.label</code>.
+     * <code>service_pattern.scheduled_stop_point_invariant.label</code>.
      */
     public final TableField<Record, String> LABEL = createField(DSL.name("label"), SQLDataType.CLOB.nullable(false), this, "");
 
@@ -57,8 +57,8 @@ public class ScheduledStopPointInvariant extends TableImpl<Record> {
 
     /**
      * Create an aliased
-     * <code>internal_service_pattern.scheduled_stop_point_invariant</code>
-     * table reference
+     * <code>service_pattern.scheduled_stop_point_invariant</code> table
+     * reference
      */
     public ScheduledStopPointInvariant(String alias) {
         this(DSL.name(alias), SCHEDULED_STOP_POINT_INVARIANT);
@@ -66,16 +66,15 @@ public class ScheduledStopPointInvariant extends TableImpl<Record> {
 
     /**
      * Create an aliased
-     * <code>internal_service_pattern.scheduled_stop_point_invariant</code>
-     * table reference
+     * <code>service_pattern.scheduled_stop_point_invariant</code> table
+     * reference
      */
     public ScheduledStopPointInvariant(Name alias) {
         this(alias, SCHEDULED_STOP_POINT_INVARIANT);
     }
 
     /**
-     * Create a
-     * <code>internal_service_pattern.scheduled_stop_point_invariant</code>
+     * Create a <code>service_pattern.scheduled_stop_point_invariant</code>
      * table reference
      */
     public ScheduledStopPointInvariant() {
@@ -88,7 +87,7 @@ public class ScheduledStopPointInvariant extends TableImpl<Record> {
 
     @Override
     public Schema getSchema() {
-        return aliased() ? null : InternalServicePattern.INTERNAL_SERVICE_PATTERN;
+        return aliased() ? null : ServicePattern.SERVICE_PATTERN;
     }
 
     @Override
